[Added]新增持久層單元測試

+++ MessageTest.Persistent.Tests/RepositoryPo/SubjectPoRepositoryTests.cs
@@ -64,24 +64,60 @@ namespace MessageTest.Persistent.Tests.RepositoryPo
             Assert.IsNotNull(addResult.subject);
             Assert.AreEqual(addResult.subject.Id, 1);
             Assert.AreEqual(addResult.subject.Title, "Test");
+
             int subjectId = 1;
             var getByIdResult = this.repo.GetById(subjectId);
-            Assert.IsNull(addResult.exception);
-            Assert.IsNotNull(addResult.subject);
-            Assert.AreEqual(addResult.subject.Id, 1);
-            Assert.AreEqual(addResult.subject.Title, "Test");
-            var deleteSubjectReqDto = new DeleteSubjectRequestDto
+            Assert.IsNull(getByIdResult.exception);
+            Assert.IsNotNull(getByIdResult.subject);
+            Assert.AreEqual(getByIdResult.subject.Id, 1);
+            Assert.AreEqual(getByIdResult.subject.Title, "Test");
+            string userId = "115051302";
+            if (getByIdResult.subject.CreatorId.Equals(userId))
+            {
+                var deleteSubjectReqDto = new DeleteSubjectRequestDto
+                {
+                    UserId = userId,
+                    Id = 1
+                };
+
+                var deleteResult = this.repo.Delete(deleteSubjectReqDto);
+
+                Assert.IsNull(deleteResult.exception);
+                Assert.IsNotNull(deleteResult.subject);
+                Assert.AreEqual(deleteResult.subject.Id, 1);
+                Assert.AreEqual(deleteResult.subject.Title, "Test");
+            }
+            else
+            {
+                Assert.IsNotNull(null);
+            }
+        }
+
+        [TestMethod]
+        public void 查詢主題留言數()
+        {
+            var addSubjectReqDto = new AddSubjectRequestDto
             {
                 UserId = "115051302",
-                Id = 1
+                SubjectTitle = "Test",
+                SubjectContent = "Test",
+                ClientTimeStamp = 1778549400,
+                CreateTimeStamp = 1778549400
             };
-
-            var delResult = this.repo.Delete(deleteSubjectReqDto);
-
+            var addResult = this.repo.Add(addSubjectReqDto);
             Assert.IsNull(addResult.exception);
             Assert.IsNotNull(addResult.subject);
             Assert.AreEqual(addResult.subject.Id, 1);
             Assert.AreEqual(addResult.subject.Title, "Test");
+
+            int subjectId = 1;
+            var getByIdResult = this.repo.GetById(subjectId);
+            Assert.IsNull(getByIdResult.exception);
+            Assert.IsNotNull(getByIdResult.subject);
+            Assert.AreEqual(getByIdResult.subject.Id, 1);
+            Assert.AreEqual(getByIdResult.subject.Title, "Test");
+            Assert.AreEqual(getByIdResult.subject.CreatorId, "115051302");
+            Assert.AreEqual(getByIdResult.subject.MessageCount, 0);
         }
     }
 }
